+++ packages/shop/locales/de.json
@@ -31,5 +31,7 @@
   "continue_browsing": "Weiter stöbern",
   "products": "Produkte",
   "contact": "Kontakt",
-  "sum": "Summe"
+  "sum": "Summe",
+  "browse": "Stöbern",
+  "continue_to_checkout": "Weiter"
 }
