+++ packages/shop/locales/en.json
@@ -31,5 +31,7 @@
   "continue_browsing": "Continue Browsing",
   "products": "Products",
   "contact": "Contact",
-  "sum": "Sum"
+  "sum": "Sum",
+  "browse": "Browse",
+  "continue_to_checkout": "Continue"
 }
